[SIMB-1846] register again on MAC change
If the main MAC address changes (because the user configured something else to
be br-lan, for example), the box would not detect it and the SIMET servers
would refuse service to an unregistered device.

Detect this by persisting the MAC registered with the SIMET servers in
/etc/config/simetbox_hash_configured, and re-register if we detect a MAC
change.

This has a better self-heal property than always using the persisted
hash_device (MAC address) from the first registering.

NOTE: this means the user has to reboot if he messes with br-lan's
underlying MAC (e.g. by reassigning it to the switch).