[SIMB-1862] Remove old version of the uptime protocol
The old simet_uptime protocol tries to do too much, and the SIMETBOX
team is really unconfortable with it.  It has also a large (>3MiB RSS)
memory footprint.

Remove it entirely, it will be replaced by a new, much safer and
simpler protocol in the future.  For now, we can use the zabbix agent
to have some idea of the uptime when absolutely required.