[MDC-155] get_mac_address: handle many exceptions
Several routers have the MAC of one of the WLAN interfaces in the label
instead of the one that ends up in br-lan.  We need to compensate for
that.

Without this, the MAC address of the SIMET probe could change, losing
previous history if it is ever factory-erased.  Also, any "defaults"
that are supposed to match the MAC address in the label, wouldn't.