[MDC-175] base: get_mac_address.sh: fix for TL-WDR3600/4300
Unfortunately, the model string was incorrect, which resulted
in TP-Link TL-WDR3600 and TL-WDR4300 using the br-lan MAC incorrectly,
instead of wlan MAC.

The incorrect MAC used will be MAC-2 in both cases.

This won't cause many issues, unless it is factory-reset or a first
install, in which case the default passwords will be MAC-2.  However,
they are shown in the web interface so user impact is minimal.