[MDC-791] simetbox-openwrt-base: change legacy hash
The "hash" in /etc/config/hash works as a device authorization token for
SIMET1, to give read permission to measurement reports, and update
permission for device metadata.

Change the hash calculation to use a device-unique invariant seed
(hopefeully, anyway), or failing that, a random number.

The old code was effectively acting always as a random number generator
(due to per-boot variable MACs such as the one for ifb0), and even if we
fixed that, it would be insecure (based only on the MAC of the
interfaces, which is information broadcast by the WiFi radios as well as
visible to network neighbors).  It also broke easily when openwrt
changed anything.

A full fleet hash reset needs to be done while deploying this change.