[MDC-791] simetbox-openwrt-base: move hash info to /etc/simet
* Use /etc/simet/simet1_hash instead of /etc/config/hash
* Use /etc/simet/simet1_deviceid instead of
  /etc/config/simetbox_hash_configured
* Add an uci-defaults script to remove /etc/config/hash and
  /etc/config/simetbox_hash_configured

We *want* to force a recalculation of the device hash (to get the new
hash), which the above acomplishes. It will also trigger the SIMET1
registration (to update the hash in the database).

This will happen only once (or as first_install), since /etc/simet/*
will be preserved by sysupgrade.