[MDC-1290] get_mac_address: better vendor API
Instead of adding to get_mac_address any vendor APIs we need, shunt to a
separate script anything that would require that vendor's firmware in
order to work.

It is fine to keep in get_mac_address any per-model behavior, though:
read vendor-only FLASH areas, or the DTS, etc: those will remain around
when running stock openwrt instead of a vendor's fork...

Output is restricted to [0-9a-z] however, with an EOL marker at the end.
[A-Z] is mapped to [a-z].  This preserves the status-quo.

The output is still pretty much a *string*, and can be of any length,
including empty.  Callers need to validate that by themselves (and
preferably just work[tm]).

Note: add as little as possible to this script.  Always prefer to fix
the DTS and send it upstream + keep a local backport.  That's the
OpenWRT way.